Just do unit test with yamale
