+++ requirements/pip.txt
@@ -64,7 +64,6 @@ Unipath==1.1
 django-kombu==0.9.4
 mimeparse==0.1.3
 mock==2.0.0
-yamale==1.7.0
 
 # stripe 1.20.2 is the latest compatible with our code base (otherwise
 # gold/tests/test_forms.py fails)
