+++ requirements/testing.txt
@@ -8,6 +8,7 @@ pytest-xdist==1.22.0
 apipkg==1.4
 execnet==1.5.0
 Mercurial==4.4.2
+yamale==1.7.0
 
 # local debugging tools
 datadiff
